+++ views/admin/en/admin_lang.php
@@ -69,4 +69,6 @@ $aLang = [
     'OSC_AMAZONPAY_REMARK'                  => 'Amazon Pay notice',
     'GENERAL_ARTICLE_OSC_AMAZON_EXCLUDE'    => 'Exclude AmazonPay',
     'OSC_AMAZONPAY_REFUND'                  => 'Refund',
+    'OSC_AMAZONPAY_REFUND_ANNOTATION'       => 'Make a refund of up to ',
+    'OSC_AMAZONPAY_DELETE_ERROR'            => 'Order cannot be deleted because money has already been transferred between the shop and Amazon.',
 ];